app/requirements.txt
fastapi
uvicorn
celery[mongodb]
boto3
pymongo
poetry
gevent
gunicorn
PyJWT
pydantic
python-multipart
